pyproject.toml
[tool.poetry]
name = "task3"
version = "0.1.0"
description = ""
authors = ["aboelsoud <mahmoudabooelsooud@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-environ = "^0.9.0"
pytz = "^2022.4"
django-model-utils = "^4.2.0"
python-dateutil = "^2.8.2"
Pillow = "^9.2.0"
django-imagekit = "^4.1.0"
setuptools = "^65.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
